+++ src/cloudflare/endpoints.rs
@@ -6,6 +6,7 @@ use crate::cloudflare::requests;
 use anyhow::{Context, Result};
 use std::collections::HashMap;
 use std::fmt::Display;
+use tracing::{debug, trace};
 
 /// Return a list of login messages if the token is verifiable.
 pub async fn verify(token: &str) -> Result<Vec<CloudfareMessage>> {
@@ -20,6 +21,7 @@ pub async fn zones(token: impl Display) -> Result<Vec<Zone>> {
     let mut zones = vec![];
     let mut page_cursor = 1;
     loop {
+        trace!("retrieving zones from page {}", page_cursor);
         let endpoint = format!("/zones?order=name&page={}", page_cursor);
         let resp: ListZonesResponse =
             requests::get(endpoint, token.to_string())
@@ -37,6 +39,7 @@ pub async fn zones(token: impl Display) -> Result<Vec<Zone>> {
             break;
         }
     }
+    debug!("collected {} zones", zones.len());
     Ok(zones)
 }
 
@@ -47,8 +50,11 @@ pub async fn records(
 ) -> Result<Vec<Record>> {
     let mut records = vec![];
     for zone in zones {
+        trace!("retrieving records from zone '{}'", zone.id);
         let mut page_cursor = 1;
+        let beginning_amt = records.len();
         loop {
+            trace!("retrieving records from page {}", page_cursor);
             let endpoint = format!(
                 "/zones/{}/dns_records?order=name&page={}",
                 zone.id, page_cursor
@@ -72,7 +78,13 @@ pub async fn records(
                 break;
             }
         }
+        debug!(
+            "received {} records from zone '{}'",
+            records.len() - beginning_amt,
+            zone.id
+        );
     }
+    debug!("collected {} records", records.len());
     Ok(records)
 }