+++ src/cmd/inventory.rs
@@ -10,12 +10,13 @@ use anyhow::{Context, Result};
 use clap::{Args, Subcommand};
 use std::{
     collections::HashSet,
-    fmt::Display,
+    fmt::{Debug, Display},
     net::{Ipv4Addr, Ipv6Addr},
-    path::{Path, PathBuf},
+    path::PathBuf,
     vec,
 };
 use tokio::time::{self, Duration, MissedTickBehavior};
+use tracing::{debug, error, info, warn};
 
 /// Build or manage your DNS record inventory.
 #[derive(Debug, Args)]
@@ -58,6 +59,7 @@ enum InventorySubcommands {
 }
 
 impl InventoryCmd {
+    #[tracing::instrument(level = "trace", skip(self, config))]
     pub async fn run(self, config: Option<PathBuf>) -> Result<()> {
         let cli_cfg = ConfigOpts {
             inventory: Some(self.cfg),
@@ -87,6 +89,7 @@ impl InventoryCmd {
     }
 }
 
+#[tracing::instrument(level = "trace", skip(opts))]
 async fn build(opts: &ConfigOpts) -> Result<()> {
     // Get token
     let token = opts
@@ -96,13 +99,12 @@ async fn build(opts: &ConfigOpts) -> Result<()> {
         .context("no token was provided")?;
 
     // Get zones and records to build inventory from
-    println!("Retrieving Cloudflare resources...");
+    info!("retrieving cloudflare resources...");
     let mut zones = cloudflare::endpoints::zones(&token).await?;
-    crate::cmd::list::filter_zones(&mut zones, opts)?;
+    crate::cmd::list::retain_zones(&mut zones, opts)?;
     anyhow::ensure!(!zones.is_empty(), "no zones to build inventory from");
-
     let mut records = cloudflare::endpoints::records(&zones, &token).await?;
-    crate::cmd::list::filter_records(&mut records, opts)?;
+    crate::cmd::list::retain_records(&mut records, opts)?;
     anyhow::ensure!(!records.is_empty(), "no records to build inventory from");
 
     let runtime = tokio::runtime::Handle::current();
@@ -157,8 +159,12 @@ async fn build(opts: &ConfigOpts) -> Result<()> {
 
             let zone_id = selected_zone.id.clone();
             let record_id = selected_record.id.clone();
-            inventory.insert(zone_id, record_id);
-            println!("✅ Added '{}'.", selected_record.name);
+            if inventory.contains(&zone_id, &record_id) {
+                println!("✅ You already added '{}'.", selected_record.name)
+            } else {
+                inventory.insert(zone_id, record_id);
+                println!("✅ Added '{}'.", selected_record.name);
+            }
         } else {
             println!("❌ No records for this zone.")
         }
@@ -197,11 +203,11 @@ async fn build(opts: &ConfigOpts) -> Result<()> {
         io::fs::remove_interactive(&path, &mut scanner).await?;
     }
     io::fs::save_yaml(&inventory, path).await?;
-    println!("✅ Saved");
 
     Ok(())
 }
 
+#[tracing::instrument(level = "trace", skip(opts))]
 async fn show(opts: &ConfigOpts) -> Result<()> {
     let inventory_path = opts
         .inventory
@@ -209,14 +215,16 @@ async fn show(opts: &ConfigOpts) -> Result<()> {
         .and_then(|opts| opts.path.clone())
         .unwrap_or_else(default_inventory_path);
     let inventory = Inventory::from_file(inventory_path).await?;
+
     if inventory.is_empty() {
-        println!("Inventory file is empty.");
+        println!("Inventory is empty.");
     } else {
         println!("{}", inventory);
     }
     Ok(())
 }
 
+#[tracing::instrument(level = "trace", skip(opts))]
 async fn check(opts: &ConfigOpts) -> Result<()> {
     // Get token
     let token = opts
@@ -226,6 +234,7 @@ async fn check(opts: &ConfigOpts) -> Result<()> {
         .context("no token was provided")?;
 
     // Get inventory
+    info!("reading inventory...");
     let inventory_path = opts
         .inventory
         .as_ref()
@@ -234,37 +243,44 @@ async fn check(opts: &ConfigOpts) -> Result<()> {
     let inventory = Inventory::from_file(inventory_path).await?;
 
     // Check records
-    println!("Checking Cloudflare resources...");
+    info!("resolving public ip...");
     let ipv4 = public_ip::addr_v4().await;
+    debug!("v4 ip: {:?}", ipv4);
     let ipv6 = public_ip::addr_v6().await;
-    let (good, bad, invalid) =
-        check_records(token, &inventory, ipv4, ipv6).await?;
+    debug!("v6 ip: {:?}", ipv4);
+    info!("retrieving cloudflare resources...");
+    let CheckResult {
+        matches,
+        mismatches,
+        invalid,
+    } = check_records(token, &inventory, ipv4, ipv6).await?;
 
     // Print records
-    for cf_record in &good {
-        println!("MATCH: {} ({})", cf_record.name, cf_record.id);
+    for cf_record in &matches {
+        info!("match: {} ({})", cf_record.name, cf_record.id);
     }
-    for cf_record in &bad {
-        println!(
-            "MISMATCH: {} ({}) => {}",
+    for cf_record in &mismatches {
+        warn!(
+            "mismatch: {} ({}) => {}",
             cf_record.name, cf_record.id, cf_record.content
         );
     }
     for (inv_zone, inv_record) in &invalid {
-        println!("INVALID: {} | {}", inv_zone, inv_record);
+        warn!("invalid: {} | {}", inv_zone, inv_record);
     }
 
     // Print summary
     println!(
-        "✅ {} GOOD, ❌ {} BAD, ❓ {} INVALID",
-        good.len(),
-        bad.len(),
+        "✅ {} matched, ❌ {} mismatched, ❓ {} invalid",
+        matches.len(),
+        mismatches.len(),
         invalid.len()
     );
 
     Ok(())
 }
 
+#[tracing::instrument(level = "trace", skip(opts))]
 async fn commit(opts: &ConfigOpts) -> Result<()> {
     // Get token
     let token = opts
@@ -274,6 +290,7 @@ async fn commit(opts: &ConfigOpts) -> Result<()> {
         .context("no token was provided")?;
 
     // Get inventory
+    info!("reading inventory...");
     let inventory_path = opts
         .inventory
         .as_ref()
@@ -286,23 +303,30 @@ async fn commit(opts: &ConfigOpts) -> Result<()> {
         .as_ref()
         .map(|opts| opts.force)
         .unwrap_or(ConfigOptsCommit::default().force);
+    debug!("force flag: {}", force);
 
     // Check records
-    println!("Checking Cloudflare resources...");
+    info!("resolving public ip...");
     let ipv4 = public_ip::addr_v4().await;
+    debug!("v4 ip: {:?}", ipv4);
     let ipv6 = public_ip::addr_v6().await;
-    let (_good, mut bad, mut invalid) =
-        check_records(&token, &inventory, ipv4, ipv6).await?;
+    debug!("v6 ip: {:?}", ipv4);
+    info!("retrieving cloudflare resources...");
+    let CheckResult {
+        mut mismatches,
+        mut invalid,
+        ..
+    } = check_records(&token, &inventory, ipv4, ipv6).await?;
 
     let runtime = tokio::runtime::Handle::current();
     let mut scanner = Scanner::new(runtime);
 
     // Print records
-    if !bad.is_empty() {
+    if !mismatches.is_empty() {
         // Print bad records
-        for cf_record in &bad {
-            println!(
-                "MISMATCH: {} ({}) => {}",
+        for cf_record in &mismatches {
+            warn!(
+                "mismatch: {} ({}) => {}",
                 cf_record.name, cf_record.id, cf_record.content
             );
         }
@@ -310,7 +334,7 @@ async fn commit(opts: &ConfigOpts) -> Result<()> {
         let fix = force
             || scanner
                 .prompt_yes_or_no(
-                    format!("Fix {} bad records?", bad.len()),
+                    format!("Fix {} mismatched records?", mismatches.len()),
                     "Y/n",
                 )
                 .await?
@@ -318,7 +342,8 @@ async fn commit(opts: &ConfigOpts) -> Result<()> {
         // Fix records
         let mut fixed = HashSet::new();
         if fix {
-            for cf_record in &bad {
+            info!("fixing mismatched records...");
+            for cf_record in &mismatches {
                 if match cf_record.record_type.as_str() {
                     "A" => match ipv4 {
                         Some(ip) => {
@@ -330,7 +355,7 @@ async fn commit(opts: &ConfigOpts) -> Result<()> {
                             )
                             .await
                         }
-                        None => Err(anyhow::anyhow!("no discovered IPv4 address needed to patch A record")),
+                        None => Err(anyhow::anyhow!("no discovered ipv4 address needed to patch A record")),
                     },
                     "AAAA" => match ipv6 {
                         Some(ip) => update_record(
@@ -340,24 +365,27 @@ async fn commit(opts: &ConfigOpts) -> Result<()> {
                                 ip,
                             )
                             .await,
-                        None => Err(anyhow::anyhow!("no discovered IPv6 address needed to patch AAAA record")),
+                        None => Err(anyhow::anyhow!("no discovered ipv6 address needed to patch AAAA record")),
                     },
                     _ => unimplemented!(
                             "unexpected record type: {}",
                             cf_record.record_type
                         ),
                 }.is_ok() {
+                    debug!("updated '{}'", &cf_record.id);
                     fixed.insert(cf_record.id.clone());
+                } else {
+                    error!("unsuccessful update of '{}'", &cf_record.id)
                 }
             }
         }
-        bad.retain_mut(|r| !fixed.contains(&r.id));
+        mismatches.retain_mut(|r| !fixed.contains(&r.id));
     }
 
     if !invalid.is_empty() {
         // Print invalid records
         for (inv_zone, inv_record) in &invalid {
-            println!("INVALID: {} | {}", inv_zone, inv_record);
+            warn!("invalid: {} | {}", inv_zone, inv_record);
         }
         // Ask to prune records
         let prune = force
@@ -371,11 +399,15 @@ async fn commit(opts: &ConfigOpts) -> Result<()> {
         // Prune
         let mut pruned = HashSet::new();
         if prune {
+            info!("pruning invalid records...");
             for (zone_id, record_id) in invalid.iter() {
                 let removed =
                     inventory.remove(zone_id.to_owned(), record_id.to_owned());
                 if let Ok(true) = removed {
+                    debug!("pruned '{} | {}'", &zone_id, &record_id);
                     pruned.insert((zone_id.clone(), record_id.clone()));
+                } else {
+                    error!("could not prune '{} | {}'", &zone_id, &record_id);
                 }
             }
             invalid.retain_mut(|(z, r)| {
@@ -386,48 +418,38 @@ async fn commit(opts: &ConfigOpts) -> Result<()> {
     }
 
     // Print summary
-    if bad.is_empty() && invalid.is_empty() {
-        println!("✅ No bad or invalid records.");
+    if mismatches.is_empty() && invalid.is_empty() {
+        info!("✅ no bad or invalid records");
     } else {
-        println!(
-            "❌ {} bad, {} invalid records remain.",
-            bad.len(),
+        info!(
+            "❌ {} mismatched, {} invalid records",
+            mismatches.len(),
             invalid.len()
         );
     }
     Ok(())
 }
 
+#[tracing::instrument(level = "trace", skip(opts))]
 pub async fn watch(opts: &ConfigOpts) -> Result<()> {
-    // Get token
-    let token = opts
-        .verify
-        .as_ref()
-        .and_then(|opts| opts.token.clone())
-        .context("no token was provided")?;
-
-    // Get inventory
-    let inventory_path = opts
-        .inventory
-        .as_ref()
-        .and_then(|opts| opts.path.clone())
-        .unwrap_or_else(default_inventory_path);
-    let mut inventory = Inventory::from_file(&inventory_path).await?;
+    // Override force flag with true; watch is non-interactive
+    let mut opts = opts.clone();
+    opts.commit.replace(ConfigOptsCommit { force: true });
 
     // Get watch interval
     let interval = Duration::from_millis(
         opts.watch
             .as_ref()
             .and_then(|opts| opts.interval)
-            .unwrap_or_else(|| ConfigOptsWatch::default().interval.unwrap()),
+            .or(ConfigOptsWatch::default().interval)
+            .context("no default interval")?,
     );
+    debug!("interval: {}ms", interval.as_millis());
 
     if interval.is_zero() {
         loop {
-            if let Err(e) =
-                __watch(&token, &mut inventory, &inventory_path).await
-            {
-                println!("Error: {:?}", e);
+            if let Err(e) = commit(&opts).await {
+                error!("{:?}", e);
             }
         }
     } else {
@@ -435,28 +457,35 @@ pub async fn watch(opts: &ConfigOpts) -> Result<()> {
         timer.set_missed_tick_behavior(MissedTickBehavior::Skip);
         loop {
             timer.tick().await;
-            if let Err(e) =
-                __watch(&token, &mut inventory, &inventory_path).await
-            {
-                println!("Error: {:?}", e);
+            debug!("awoken");
+            if let Err(e) = commit(&opts).await {
+                error!("{:?}", e);
             }
-            println!("Sleeping...");
+            debug!("sleeping...");
         }
     }
 }
 
+#[derive(Debug, Clone)]
+pub struct CheckResult {
+    matches: Vec<Record>,
+    mismatches: Vec<Record>,
+    invalid: Vec<(String, String)>,
+}
+
+#[tracing::instrument(level = "trace")]
 pub async fn check_records(
-    token: impl Display,
+    token: impl Display + Debug,
     inventory: &Inventory,
     ipv4: Option<Ipv4Addr>,
     ipv6: Option<Ipv6Addr>,
-) -> Result<(Vec<Record>, Vec<Record>, Vec<(String, String)>)> {
+) -> Result<CheckResult> {
     let zones = cloudflare::endpoints::zones(token.to_string()).await?;
     let records =
         cloudflare::endpoints::records(&zones, token.to_string()).await?;
 
     // Check and collect records
-    let (mut good, mut bad, mut invalid) = (vec![], vec![], vec![]);
+    let (mut matches, mut mismatches, mut invalid) = (vec![], vec![], vec![]);
     for (inv_zone, inv_records) in inventory.clone().into_iter() {
         for inv_record in inv_records {
             let cf_record = records.iter().find(|r| {
@@ -476,10 +505,10 @@ pub async fn check_records(
                     if let Some(ref ip) = ip {
                         if &cf_record.content == ip {
                             // IP Match
-                            good.push(cf_record.clone());
+                            matches.push(cf_record.clone());
                         } else {
                             // IP mismatch
-                            bad.push(cf_record.clone());
+                            mismatches.push(cf_record.clone());
                         }
                     } else {
                         anyhow::bail!(
@@ -496,101 +525,9 @@ pub async fn check_records(
         }
     }
 
-    Ok((good, bad, invalid))
-}
-
-/// This would be fantastic as an async closure when that becomes stabalized.
-/// For now, this is a helper to perform the commits without interaction.
-async fn __watch<P>(
-    token: impl Display,
-    inventory: &mut Inventory,
-    inventory_path: P,
-) -> Result<()>
-where
-    P: AsRef<Path>,
-{
-    let token = token.to_string();
-
-    // Check records
-    println!("Checking Cloudflare resources...");
-    let ipv4 = public_ip::addr_v4().await;
-    let ipv6 = public_ip::addr_v6().await;
-    let (_good, mut bad, mut invalid) =
-        check_records(&token, inventory, ipv4, ipv6).await?;
-
-    // Print records
-    if !bad.is_empty() {
-        // Fix records
-        let mut fixed = HashSet::new();
-        for cf_record in &bad {
-            println!(
-                "MISMATCH: {} ({}) => {}",
-                cf_record.name, cf_record.id, cf_record.content
-            );
-            if match cf_record.record_type.as_str() {
-                    "A" => match ipv4 {
-                        Some(ip) => {
-                            update_record(
-                                token.clone(),
-                                cf_record.zone_id.clone(),
-                                cf_record.id.clone(),
-                                ip,
-                            )
-                            .await
-                        }
-                        None => Err(anyhow::anyhow!("no discovered IPv4 address needed to patch A record")),
-                    },
-                    "AAAA" => match ipv6 {
-                        Some(ip) => update_record(
-                                token.clone(),
-                                cf_record.zone_id.clone(),
-                                cf_record.id.clone(),
-                                ip,
-                            )
-                            .await,
-                        None => Err(anyhow::anyhow!("no discovered IPv6 address needed to patch AAAA record")),
-                    },
-                    _ => unimplemented!(
-                            "unexpected record type: {}",
-                            cf_record.record_type
-                        ),
-                }.is_ok() {
-                    fixed.insert(cf_record.id.clone());
-                }
-        }
-        bad.retain_mut(|r| !fixed.contains(&r.id));
-    }
-
-    // Prune
-    let mut pruned = HashSet::new();
-    if !invalid.is_empty() {
-        // Print invalid records
-        for (inv_zone, inv_record) in &invalid {
-            println!("INVALID: {} | {}", inv_zone, inv_record);
-        }
-        for (zone_id, record_id) in invalid.iter() {
-            let removed =
-                inventory.remove(zone_id.to_owned(), record_id.to_owned());
-            if let Ok(true) = removed {
-                pruned.insert((zone_id.clone(), record_id.clone()));
-            }
-        }
-        invalid.retain_mut(|(z, r)| {
-            !pruned.contains(&(z.to_owned(), r.to_owned()))
-        });
-        inventory.save(inventory_path).await?;
-    }
-
-    // Print summary
-    if bad.is_empty() && invalid.is_empty() {
-        println!("✅ No bad or invalid records.");
-    } else {
-        println!(
-            "❌ {} bad, {} invalid records remain.",
-            bad.len(),
-            invalid.len()
-        );
-    }
-
-    Ok(())
+    Ok(CheckResult {
+        matches,
+        mismatches,
+        invalid,
+    })
 }