+++ src/cmd/verify.rs
@@ -5,6 +5,7 @@ use crate::{
 use anyhow::{Context, Result};
 use clap::Args;
 use std::path::PathBuf;
+use tracing::info;
 
 /// Verify authentication to Cloudflare.
 #[derive(Debug, Args)]
@@ -15,6 +16,7 @@ pub struct VerifyCmd {
 }
 
 impl VerifyCmd {
+    #[tracing::instrument(level = "trace", skip(self, config))]
     pub async fn run(self, config: Option<PathBuf>) -> Result<()> {
         let cli_cfg = ConfigOpts {
             verify: Some(self.cfg),
@@ -22,22 +24,24 @@ impl VerifyCmd {
         };
         let opts = ConfigOpts::full(config, Some(cli_cfg))?;
 
-        // Get token
-        let token = opts
-            .verify
-            .as_ref()
-            .and_then(|opts| opts.token.clone())
-            .context("no token was provided")?;
-
-        println!("Verifying...");
-        let login_messages = cloudflare::endpoints::verify(&token).await?;
-        if let Some(message_stack) = login_messages
-            .into_iter()
-            .map(|msg| msg.message)
-            .reduce(|cur: String, nxt: String| cur + "\n" + &nxt)
-        {
-            println!("{}", message_stack);
-        }
-        Ok(())
+        verify(&opts).await
     }
+}
+
+#[tracing::instrument(level = "trace", skip(opts))]
+async fn verify(opts: &ConfigOpts) -> Result<()> {
+    // Get token
+    let token = opts
+        .verify
+        .as_ref()
+        .and_then(|opts| opts.token.clone())
+        .context("no token was provided")?;
+
+    info!("validating...");
+    let cf_messages = cloudflare::endpoints::verify(&token).await?;
+    info!("received successful response");
+    for (i, msg) in cf_messages.iter().enumerate() {
+        info!("message [{}/{}]: {}", i + 1, cf_messages.len(), msg.message);
+    }
+    Ok(())
 }